Refactor Kafka admin usage
